增加Compiler等

+++ PomeloLanguage/PomeloLanguage/Util/Unicode.swift
@@ -8,7 +8,7 @@
 
 import Cocoa
 
-typealias Byte = UInt8
+typealias Byte = UInt64
 class Unicode {
     public static func getByteOfEncodeUtf8(value: Int) -> Int {
         assert(value > 0)
